remastered project structure
