+++ .gitignore
@@ -1,5 +1,19 @@
-copy
+/target/
+!.mvn/wrapper/maven-wrapper.jar
+
+### IntelliJ IDEA ###
 .idea
-target
+*.iws
 *.iml
+*.ipr
+.mvn
 *.jar
+*.war
+*.nar
+*.ear
+*.zip
+*.tar.gz
+*.rar
+
+# Compiled class file
+*.class
