--- src/cli/pom.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>cli</artifactId>
-    <version>1.0</version>
-    <packaging>jar</packaging>
-    <name>cli</name>
-
-    <parent>
-        <groupId>crossgen</groupId>
-        <artifactId>cross-gen</artifactId>
-        <version>1.0</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <dependencies>
-        <dependency>
-            <groupId>crossgen</groupId>
-            <artifactId>game</artifactId>
-            <version>1.0</version>
-        </dependency>
-    </dependencies>
-</project>
