--- src/cli/src/main/java/crossgen/cli/CommandDispatcher.java
@@ -1,7 +1,0 @@
-package crossgen.cli;
-
-public interface CommandDispatcher {
-    //boolean isCommand(int commandNumber);
-    void executeCommand(int commandNumber) throws Exception;
-    void addCommand(Command command);
-}
