--- src/logic/src/main/java/crossgen/logic/generator/Generator.java
@@ -1,7 +1,0 @@
-package crossgen.logic.generator;
-
-import crossgen.logic.grid.Grid;
-
-public interface Generator {
-    Grid generate();
-}
