+++ src/main/java/netcracker/practice/crossgen/Main.java
@@ -1,0 +1,9 @@
+package netcracker.practice.crossgen;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+    }
+
+}
