+++ src/main/java/netcracker/practice/crossgen/cli/CommandDispatcher.java
@@ -1,0 +1,7 @@
+package netcracker.practice.crossgen.cli;
+
+public interface CommandDispatcher {
+    //boolean isCommand(int commandNumber);
+    void executeCommand(int commandNumber) throws Exception;
+    void addCommand(Command command);
+}
