+++ src/main/java/netcracker/practice/crossgen/game/CrosswordGame.java
@@ -1,0 +1,48 @@
+package netcracker.practice.crossgen.game;
+
+import netcracker.practice.crossgen.logic.crossword.CanadianCrossword;
+
+public class CrosswordGame implements Game {
+    private CanadianCrossword crossword;
+    private GameState state;
+
+    public CrosswordGame() {
+        state = new InitialState();
+    }
+
+    public void play() {
+        // TODO : implement game methods
+    }
+
+    public void configure() {
+
+    }
+
+    public void getHelp() {
+
+    }
+
+    public void quit() {
+
+    }
+
+    public void delete() {
+
+    }
+
+    public void create() {
+
+    }
+
+    public void save() {
+
+    }
+
+    public void check() {
+
+    }
+
+    public void handleInput(String input) {
+
+    }
+}
