+++ src/main/java/netcracker/practice/crossgen/game/Game.java
@@ -1,0 +1,13 @@
+package netcracker.practice.crossgen.game;
+
+public interface Game {
+    void play();
+    void configure();
+    void getHelp();
+    void quit();
+    void delete();
+    void create();
+    void save();
+    void check();
+    void handleInput(String input);
+}
