+++ src/main/java/netcracker/practice/crossgen/logic/crossword/CanadianCrossword.java
@@ -1,0 +1,22 @@
+package netcracker.practice.crossgen.logic.crossword;
+
+import java.util.List;
+import java.util.ArrayList;
+
+import netcracker.practice.crossgen.logic.grid.Grid;
+
+public class CanadianCrossword extends Grid {
+    private List<Clue> clues = new ArrayList<>();
+
+    public CanadianCrossword(int width, int height) {
+        super(width, height);
+    }
+
+    public List<Clue> getClues() {
+        return this.clues;
+    }
+
+    public void addClue(Clue clue) {
+        clues.add(clue);
+    }
+}
