+++ src/main/java/netcracker/practice/crossgen/logic/generator/Generator.java
@@ -1,0 +1,9 @@
+package netcracker.practice.crossgen.logic.generator;
+
+import netcracker.practice.crossgen.logic.grid.Grid;
+
+public interface Generator {
+
+    Grid generate();
+
+}
