+++ src/main/java/netcracker/practice/crossgen/logic/grid/Angle.java
@@ -1,0 +1,10 @@
+package netcracker.practice.crossgen.logic.grid;
+
+import java.util.*;
+
+public interface Angle {
+
+    List<Word> findGridWords(Grid grid);
+    Map<Word, Map<Word, Integer>> findGridIntersections(Grid grid);
+
+}
