+++ src/main/java/netcracker/practice/crossgen/logic/grid/AnglePicker.java
@@ -1,0 +1,15 @@
+package netcracker.practice.crossgen.logic.grid;
+
+public class AnglePicker {
+    public Angle getAngle(String angleType) {
+        if (angleType == null)
+            return null;
+        if(angleType.equalsIgnoreCase("straight")){
+            return new StraightAngle();
+
+        } else if(angleType.equalsIgnoreCase("diagonal")) {
+            return new DiagonalAngle();
+        }
+        return null;
+    }
+}
