+++ src/main/java/netcracker/practice/crossgen/logic/grid/DiagonalAngle.java
@@ -1,0 +1,26 @@
+package netcracker.practice.crossgen.logic.grid;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+
+public class DiagonalAngle implements Angle {
+
+    @Override
+    public List<Word> findGridWords(Grid grid) {
+        List<Word> gridWords = new ArrayList<>();
+
+        // TODO : implement diagonal strategy
+
+        return gridWords;
+    }
+
+    @Override
+    public Map<Word, Map<Word, Integer>> findGridIntersections(Grid grid) {
+
+        // TODO : implement diagonal strategy
+
+        return null;
+    }
+
+}
