--- src/main/pom.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>main</artifactId>
-    <version>1.0</version>
-    <packaging>jar</packaging>
-    <name>main</name>
-
-    <parent>
-        <groupId>crossgen</groupId>
-        <artifactId>cross-gen</artifactId>
-        <version>1.0</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <build>
-        <pluginManagement>
-            <plugins>
-                <plugin>
-                    <artifactId>maven-jar-plugin</artifactId>
-                    <version>3.0.2</version>
-                    <configuration>
-                        <archive>
-                            <addMavenDescriptor>false</addMavenDescriptor>
-                            <compress>true</compress>
-                            <manifest>
-                                <addClasspath>true</addClasspath>
-                                <classpathPrefix>libs/</classpathPrefix>
-                                <mainClass>crossgen.main.Main</mainClass>
-                            </manifest>
-                        </archive>
-                    </configuration>
-                </plugin>
-                <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-dependency-plugin</artifactId>
-                    <version>3.1.1</version>
-                    <executions>
-                        <execution>
-                            <id>copy-dependencies</id>
-                            <phase>package</phase>
-                            <goals>
-                                <goal>copy-dependencies</goal>
-                            </goals>
-                            <configuration>
-                                <outputDirectory>${project.build.directory}/libs</outputDirectory>
-                            </configuration>
-                        </execution>
-                    </executions>
-                </plugin>
-            </plugins>
-        </pluginManagement>
-    </build>
-
-
-    <dependencies>
-        <dependency>
-            <groupId>crossgen</groupId>
-            <artifactId>cli</artifactId>
-            <version>1.0</version>
-        </dependency>
-        <dependency>
-            <groupId>crossgen</groupId>
-            <artifactId>ui</artifactId>
-            <version>1.0</version>
-        </dependency>
-        <dependency>
-            <groupId>crossgen</groupId>
-            <artifactId>game</artifactId>
-            <version>1.0</version>
-        </dependency>
-    </dependencies>
-</project>
