+++ src/test/java/netcracker/practice/crossgen/MainTest.java
@@ -1,0 +1,8 @@
+package netcracker.practice.crossgen;
+
+import static org.junit.Assert.assertTrue;
+import org.junit.Test;
+
+public class MainTest {
+
+}
